+++ include/tensorpack.h
@@ -40,13 +40,13 @@ class TensorPack
         void add_tensor(const std::string& name /*!< The name used to reference the tensor*/,
                         const std::string& type /*!< The data type of the tensor*/,
                         void* data /*!< A c_ptr to the data of the tensor*/,
-                        const std::vector<int>& dims /*! The dimensions of the data*/
+                        const std::vector<size_t>& dims /*! The dimensions of the data*/
                         );
 
         //! Method for adding a tensor using a buffer instead of data
         void add_tensor(const std::string& name /*!< The name used to reference the tensor*/,
                         const std::string& type /*!< The data type of the tensor*/,
-                        const std::vector<int>& dims /*! The dimensions of the data*/,
+                        const std::vector<size_t>& dims /*! The dimensions of the data*/,
                         const std::string_view& buf /*!< A c_ptr to the data of the tensor*/
                         );
